+++ packages/ispectify/lib/ispectify.dart
@@ -13,6 +13,7 @@ export 'src/models/exception.dart';
 export 'src/models/models.dart';
 export 'src/observer.dart';
 export 'src/options.dart';
+export 'src/redaction/redaction_service.dart';
 export 'src/settings.dart';
 export 'src/truncator.dart';
 export 'src/utils/console_utils.dart';