+++ packages/ispectify_http/lib/src/models/response.dart
@@ -16,11 +16,12 @@ class HttpResponseLog extends ISpectifyData {
     required this.requestBody,
     required this.responseBody,
     required this.settings,
+    this.redactor,
   }) : super(
           key: getKey,
           title: getKey,
           pen: settings.responsePen ?? (AnsiPen()..xterm(35)),
-          additionalData: responseData?.toJson,
+          additionalData: responseData?.toJson(redactor: redactor),
         );
 
   final String? method;
@@ -34,6 +35,7 @@ class HttpResponseLog extends ISpectifyData {
   final Object? responseBody;
   final ISpectHttpInterceptorSettings settings;
   final HttpResponseData? responseData;
+  final RedactionService? redactor;
 
   static const getKey = 'http-response';